third_party/blink/web_tests/external/wpt/cookie-store/cookieStore_set_limit.https.any-expected.txt

This is a testharness.js-based test.
[FAIL] Set max-size value-less cookie
  assert_equals: expected (undefined) undefined but got (object) object "[object Object]"
Harness: the test ran to completion.
